chore:  support `testing.T#Cleanup()` method (#1)
## Summary

This pull request introduces a wrapper for `*rapid.T` called `rapidT` to
address compatibility issues with the `TestingT` interface. The changes
include adding a `Cleanup()` method to the `TestingT` interface and
implementing the `rapidT` wrapper with a stub for the `Cleanup()`
method. The wrapper is used in `runScenario` function to maintain
compatibility and provide better integration between `rapid` and
`testing` packages.

*NOTE: step definition signatures using `*rapid.T` are unaffected (i.e.
backwards compatible change)*

## Rationale

We have some non-trivial, common test helpers that occasionally reach
for `t.Cleanup(...)`. I acknowledge that the `After` is intended for
suite cleanup but I think this is an equally valid case.